refactor: consolidate to one CDK crate
Having the two crates adds complexity without
benefit since features can be used instead
