+++ Cargo.toml
@@ -1,6 +1,5 @@
 [workspace]
 members = [
-    "crates/cashu",
     "crates/cdk",
 ]
 resolver = "2"
@@ -21,15 +20,7 @@ license-file = "LICENSE"
 keywords = ["bitcoin", "e-cash", "cashu"]
 
 [workspace.dependencies]
-serde = { version = "1.0.160", features = ["derive"]}
-serde_json = "1.0.96"
-url = "2.3.1"
 tokio = { version = "1.32", default-features = false }
-tracing = { version = "0.1", default-features = false }
-tracing-subscriber = "0.3"
-uniffi = "0.24"
-thiserror = "1.0.50"
-getrandom = { version = "0.2", features = ["js"] }
 
 [profile]
