--- crates/cashu/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "cashu"
-version = "0.4.1-ALPHA"
-edition = "2021"
-authors = ["thesimplekid"]
-readme = "README.md"
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-rust-version.workspace = true # MSRV
-description = "Cashu rust wallet and mint library"
-
-
-[features]
-default = ["mint", "wallet", "all-nuts"]
-mint = []
-wallet = []
-all-nuts = ["nut13"]
-nut13 = ["dep:bip39"]
-
-[dependencies]
-base64 = "0.21" # bitcoin uses v0.21 (optional dep)
-bip39 = { version = "2.0", optional = true }
-bitcoin = { version = "0.30", features = ["serde", "rand", "rand-std"] } # lightning-invoice uses v0.30
-lightning-invoice = { version = "0.29", features = ["serde"] }
-once_cell = "1.19"
-serde.workspace = true
-serde_json.workspace = true
-serde_with = "3.4"
-url.workspace = true
-thiserror.workspace = true
-tracing.workspace = true
-uuid = { version = "1.6", features = ["v4"] }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-getrandom.workspace = true
-instant = { version = "0.1", features = [ "wasm-bindgen", "inaccurate" ] }
