+++ crates/cdk/Cargo.toml
@@ -11,25 +11,30 @@ license.workspace = true
 
 [features]
 default = ["mint", "wallet", "all-nuts", "redb"]
-mint = ["cashu/mint"]
-wallet = ["cashu/wallet", "dep:minreq"]
+mint = []
+wallet = ["dep:minreq"]
 gloo = ["dep:gloo"]
 all-nuts = ["nut13"]
-nut13 = ["cashu/nut13"]
+nut13 = ["dep:bip39"]
 redb = ["dep:redb"]
 
 
 [dependencies]
-bip39 = "2.0.0"
-cashu = { path = "../cashu" }
-serde = { workspace = true }
-serde_json = { workspace = true }
-url = { workspace = true }
-tracing = { workspace = true }
-thiserror = { workspace = true }
+base64 = "0.21" # bitcoin uses v0.21 (optional dep)
+bip39 = { version = "2.0", optional = true }
+bitcoin = { version = "0.30", features = ["serde", "rand", "rand-std"] } # lightning-invoice uses v0.30
+lightning-invoice = { version = "0.29", features = ["serde"] }
+once_cell = "1.19"
+serde = { version = "1.0.160", default-features = false, features = ["derive"]}
+serde_json = { version = "1.0.96", default-features = false }
+serde_with = "3.4"
+url = "2.3.1"
+tracing = { version = "0.1", default-features = false }
+thiserror = "1.0.50"
 async-trait = "0.1.74"
 gloo = { version = "0.11.0", optional = true, features = ["net"] }
 http = "1.0.0"
+uuid = { version = "1.6", features = ["v4"] }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 tokio = { workspace = true, features = ["rt-multi-thread", "time", "macros", "sync"] }
@@ -38,6 +43,7 @@ redb = { version = "2.0.0", optional = true }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 tokio = { workspace = true, features = ["rt", "macros", "sync", "time"] }
-getrandom = { workspace = true }
+getrandom = { version = "0.2.14" }
+instant = { version = "0.1", features = [ "wasm-bindgen", "inaccurate" ] }
