+++ crates/cdk/src/nuts/nut00.rs
