+++ crates/cdk/src/nuts/nut02.rs
