+++ crates/cdk/src/nuts/nut06.rs
